Move `h.views.api_users` -> `h.views.api.users`
